+++ lib/src/ast/sass/expression/binary_operation.dart
@@ -6,6 +6,7 @@ import 'package:charcode/charcode.dart';
 import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
+import '../../../util/span.dart';
 import '../../../visitor/interface/expression.dart';
 import '../expression.dart';
 import 'list.dart';
@@ -44,6 +45,17 @@ final class BinaryOperationExpression implements Expression {
     }
     return left.span.expand(right.span);
   }
+
+  /// Returns the span that covers only [operator].
+  ///
+  /// @nodoc
+  @internal
+  FileSpan get operatorSpan => left.span.file == right.span.file &&
+          left.span.end.offset < right.span.start.offset
+      ? left.span.file
+          .span(left.span.end.offset, right.span.start.offset)
+          .trim()
+      : span;
 
   BinaryOperationExpression(this.operator, this.left, this.right)
       : allowsSlash = false;