+++ pkg/sass_api/lib/sass_api.dart
@@ -23,6 +23,7 @@ export 'package:sass/src/visitor/find_dependencies.dart';
 export 'package:sass/src/visitor/interface/expression.dart';
 export 'package:sass/src/visitor/interface/selector.dart';
 export 'package:sass/src/visitor/interface/statement.dart';
+export 'package:sass/src/visitor/ast_search.dart';
 export 'package:sass/src/visitor/recursive_ast.dart';
 export 'package:sass/src/visitor/recursive_selector.dart';
 export 'package:sass/src/visitor/recursive_statement.dart';